backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt
email-validator==2.1.0
APScheduler==3.10.4
